Probably done here.
I have ran into an error testing out the client code using the server to email. After looking up the issue (code: 'EAUTH', command: 'AUTH XOAUTH2'), I find Gmail has high security and it is an unavoidable issue at this time.

The code for the form is contacting the api middleware here and it should be getting back a successful response but tis not the case.